+++ R/decay_binary.R
@@ -5,7 +5,8 @@
 #' functions.
 #' @template description_generic_cost
 #'
-#' @param cutoff A `numeric`. A number indicating the travel cost cutoff.
+#' @param cutoff A `numeric` vector. The numbers indicating the travel cost
+#'   cutoffs.
 #'
 #' @template return_decay_function
 #'
@@ -14,17 +15,34 @@
 #' @examples
 #' weighting_function <- decay_binary(cutoff = 30)
 #'
-#' weighting_function(20)
+#' weighting_function(c(20, 35))
 #'
-#' weighting_function(35)
+#' weighting_function <- decay_binary(cutoff = c(30, 45))
+#'
+#' weighting_function(c(20, 35))
 #'
 #' @export
 decay_binary <- function(cutoff) {
-  checkmate::assert_number(cutoff, lower = 0, finite = TRUE)
+  checkmate::assert_numeric(
+    cutoff,
+    lower = 0,
+    finite = TRUE,
+    any.missing = FALSE,
+    min.len = 1,
+    unique = TRUE
+  )
 
   weighting_function <- function(travel_cost) {
-    weights <- as.integer(travel_cost <= cutoff)
-    return(weights)
+    weights_list <- lapply(
+      cutoff,
+      function(x) {
+        weights <- as.integer(travel_cost <= x)
+        weights
+      }
+    )
+    names(weights_list) <- cutoff
+
+    return(weights_list)
   }
 
   return(weighting_function)